+++ phpBB2/ctracker/admin/acp_module_changedfiles.php
@@ -123,9 +123,13 @@ $template->set_filenames(array(
 /*
  * Send some vars to the template
  */
+$last_checksum_scan = intval($ctracker_config->settings['last_checksum_scan']);
+$last_checksum_scan_text = ($last_checksum_scan > 0)
+	? date($board_config['default_dateformat'], $last_checksum_scan)
+	: $lang['ctracker_never'];
 $template->assign_vars(array(
 	'L_HEADLINE' 		=> $lang['ctracker_fchk_head'],
-	'L_SUBHEADLINE'		=> sprintf($lang['ctracker_fchk_subhead'], date($board_config['default_dateformat'], $ctracker_config->settings['last_checksum_scan'])),
+	'L_SUBHEADLINE'		=> sprintf($lang['ctracker_fchk_subhead'], $last_checksum_scan_text),
 	'L_FUNC_HEADER'		=> $lang['ctracker_fchk_funcheader'],
 	'L_TABLE_HEADER'	=> $lang['ctracker_fchk_tableheader'],
 	'L_OPTION_1'		=> $lang['ctracker_fchk_option1'],